requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
openpyxl>=3.0.0
XlsxWriter>=3.0.0
yagmail>=0.15.0
fpdf2>=1.7.0
plotly>=5.0.0
python-dateutil>=2.8.0